refactor: Split join reorder.rs into smaller, focused modules
Breaks down the 838-line reorder.rs into 5 smaller modules for better
maintainability and separation of concerns:

- **predicates.rs** (237 lines) - Predicate extraction and analysis
  - extract_table_local_predicates: Identifies single-table predicates
  - extract_in_predicates_from_or: TPC-H Q7 optimization
  - extract_where_equijoins: Finds join conditions in WHERE clauses
  - flatten_and_chain: Breaks down AND expressions

- **graph.rs** (221 lines) - Join graph construction and table analysis
  - TableRef struct: Table metadata and references
  - flatten_join_tree: Converts nested joins to flat table list
  - extract_all_conditions: Collects join conditions
  - extract_referenced_tables: Determines table involvement

- **utils.rs** (176 lines) - Utility functions and validation
  - should_apply_join_reordering: Determines when to optimize
  - count_tables_in_from: Counts tables in FROM clause
  - all_joins_are_cross: Validates join types
  - build_reordered_schema: Restores original column ordering
  - build_column_permutation: Maps optimized to original positions

- **optimizer.rs** (245 lines) - Main optimization logic
  - execute_with_join_reordering: Coordinates optimization pipeline
  - Implements cost-based join order search
  - Executes joins in optimal order
  - Restores original column semantics

- **mod.rs** (24 lines) - Public API and module coordination

Benefits:
- Reduced cognitive load when working on specific optimization aspects
- Clear separation between predicate analysis, graph construction, and optimization
- Easier to test and maintain individual components
- Better encapsulation with module-level visibility controls
- No changes to public API - maintains backward compatibility

Closes #2460
